requirements_desktop.txt
fastapi>=0.110
uvicorn[standard]>=0.23
sqlalchemy>=2.0
pydantic>=2.4
openpyxl>=3.1
bcrypt>=4.0.1
python-multipart>=0.0.6
python-dateutil>=2.8.2
pywebview>=4.0
pyinstaller>=6.0
